+++ src/ml/mcos/liteteleport/config/HomeInfo.java
@@ -45,7 +45,7 @@ public class HomeInfo {
         Config.saveConfiguration(homeInfo, homes);
     }
 
-    public static void setHome(String player, String homeName, String world, double x, double y, double z, float yaw, float pitch) {
+    public static void setHome(String player, String homeName, String world, double x, double y, double z, double yaw, double pitch) {
         Config.setLocation(homeInfo, player + "." + escape(homeName), world, x, y, z, yaw, pitch);
     }
